requirements.txt
# Core Dependencies
mistralai>=0.1.0
transformers>=4.36.0
torch>=2.1.0
sentence-transformers>=2.2.2

# Biomedical NLP
biopython>=1.81
scispacy>=0.5.3
spacy>=3.7.0

# Data & Vector Store
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0

# Web & API
requests>=2.31.0
beautifulsoup4>=4.12.0
streamlit>=1.29.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
networkx>=3.2.0

# Testing & Quality
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0

# Config & Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.2
